refactor: :fire: join all user entities in user package
Join user_benefit_purchase and user_good_action in user, because all is about user
